added delete confirm

+++ root/home.php
@@ -29,9 +29,9 @@ $result_courses = mysql_query($query_courses) or die(mysql_error());
         <link rel="stylesheet" type="text/css" href="css/data_table.css" />
         <link rel="stylesheet" type="text/css" href="css/jquery.ui.theme.css" />
         <link rel="stylesheet" type="text/css" href="css/ui_darkness.css" />
-        
+
         <link rel="stylesheet" type="text/css" href="engine1/style.css" />
-	<script type="text/javascript" src="engine1/jquery.js"></script>
+        <script type="text/javascript" src="engine1/jquery.js"></script>
 
         <script src="js/jquery-1.7.2.js" type="text/javascript"></script>
         <script src="js/jquery.dataTables.min.js" type="text/javascript"></script>
@@ -52,7 +52,7 @@ $result_courses = mysql_query($query_courses) or die(mysql_error());
                 <!-- end .header -->
                 <?php
                 // Displaying message and errors
-                include 'info.php';
+                include '../includes/info.php';
                 ?>
             </div>
             <div class="content">
@@ -173,7 +173,7 @@ $result_courses = mysql_query($query_courses) or die(mysql_error());
                                         echo '<td>' . $row_news['nws_description'] . '</td>';
                                         echo '<td><a href="uploads/docs/' . $row_news['nws_attachment'] . '">' . $row_news['nws_attachment'] . '</a></td>';
                                         echo '<td><a href="edit_news.php?id=' . $row_news['nws_id'] . '" class="edit-news">Edit</a></td>';
-                                        echo '<td><a href="delete_news.php?id=' . $row_news['nws_id'] . '">Delete</a></td>';
+                                        echo '<td><a href="delete_news.php?id=' . $row_news['nws_id'] . '" onClick="return confirm(\'Are you sure you want to delete this news\');">Delete</a></td>';
                                         echo '</tr>';
                                     }
                                     ?>
@@ -206,7 +206,7 @@ $result_courses = mysql_query($query_courses) or die(mysql_error());
                                         echo '<td>' . $row_evnt['event_description'] . '</td>';
                                         echo '<td><a href="uploads/docs/' . $row_evnt['event_attachment'] . '">' . $row_evnt['event_attachment'] . '</a></td>';
                                         echo '<td><a href="edit_events.php?id=' . $row_evnt['event_id'] . '" class="edit-events">Edit</a></td>';
-                                        echo '<td><a href="delete_events.php?id=' . $row_evnt['event_id'] . '">Delete</a></td>';
+                                        echo '<td><a href="delete_events.php?id=' . $row_evnt['event_id'] . '" onClick="return confirm(\'Are you sure you want to delete this event\');">Delete</a></td>';
                                         echo '</tr>';
                                     }
                                     ?>
@@ -236,7 +236,7 @@ $result_courses = mysql_query($query_courses) or die(mysql_error());
                                         echo '<td>' . $row_courses['course_posted_date'] . '</td>';
                                         echo '<td>' . $row_courses['course_description'] . '</td>';
                                         echo '<td><a href="edit_courses.php?id=' . $row_courses['course_id'] . '" class="edit-courses">Edit</a></td>';
-                                        echo '<td><a href="delete_courses.php?id=' . $row_courses['course_id'] . '">Delete</a></td>';
+                                        echo '<td><a href="delete_courses.php?id=' . $row_courses['course_id'] . '" onClick="return confirm(\'Are you sure you want to delete this course\');">Delete</a></td>';
                                         echo '</tr>';
                                     }
                                     ?>
